basic verifier test

+++ src/BasicVerifier.sol
@@ -44,15 +44,21 @@ contract BasicVerifier {
 
     function domain() internal pure returns (Radix2EvaluationDomain memory) {
         return Radix2EvaluationDomain({
-           size: 256,
-           log_size_of_group: LOG_N,
-           size_as_field_element: Bw6Fr(0, 256),
-           size_inv: Bw6Fr(0x1ac8c0bd1ad4bd9db74cabaac34a7f1, 0xdf08b7190df41e7b8fd46ecd8a4f3eb816f451e6ebd000008483b74000000001),
-           group_gen: Bw6Fr(0x1002f29b90f34d050aca1e5fa3f7633, 0x712d6bc0d484c501aed11a0c88d9f8c87a0c14230db0b91cb42b84a2dce33f04),
-           group_gen_inv: Bw6Fr(0x3c41de81751c63b6e0389795b5feba, 0xeffa4f9e8f3b8a87ee9ee3eabaa65f2ab40a45e807458704a002ea0add4e93c5),
-           offset: BW6FR.one(),
-           offset_inv: BW6FR.one(),
-           offset_pow_size: BW6FR.one()
+            size: 256,
+            log_size_of_group: LOG_N,
+            size_as_field_element: Bw6Fr(0, 256),
+            size_inv: Bw6Fr(
+                0x1ac8c0bd1ad4bd9db74cabaac34a7f1, 0xdf08b7190df41e7b8fd46ecd8a4f3eb816f451e6ebd000008483b74000000001
+                ),
+            group_gen: Bw6Fr(
+                0x1002f29b90f34d050aca1e5fa3f7633, 0x712d6bc0d484c501aed11a0c88d9f8c87a0c14230db0b91cb42b84a2dce33f04
+                ),
+            group_gen_inv: Bw6Fr(
+                0x3c41de81751c63b6e0389795b5feba, 0xeffa4f9e8f3b8a87ee9ee3eabaa65f2ab40a45e807458704a002ea0add4e93c5
+                ),
+            offset: BW6FR.one(),
+            offset_inv: BW6FR.one(),
+            offset_pow_size: BW6FR.one()
         });
     }
 
@@ -62,7 +68,7 @@ contract BasicVerifier {
         Bls12G2 calldata aggregate_signature,
         KeysetCommitment calldata new_validator_set_commitment
     ) external view {
-        uint n_signers = public_input.bitmask.count_ones();
+        uint256 n_signers = public_input.bitmask.count_ones();
         // apk proof verification
         require(verify_simple(public_input, proof), "!apk");
         // aggregate BLS signature verification
@@ -81,10 +87,11 @@ contract BasicVerifier {
         return BLS12Pairing.verify(aggregate_public_key, aggregate_signature, message);
     }
 
-    function verify_simple(
-        AccountablePublicInput calldata public_input,
-        SimpleProof calldata proof
-    ) internal view returns (bool) {
+    function verify_simple(AccountablePublicInput calldata public_input, SimpleProof calldata proof)
+        internal
+        view
+        returns (bool)
+    {
         Challenges memory challenges = restore_challenges(public_input, proof, POLYS_OPENED_AT_ZETA);
         LagrangeEvaluations memory evals_at_zeta = challenges.zeta.lagrange_evaluations(domain());
         Bw6Fr memory b_at_zeta = challenges.zeta.barycentric_eval_binary_at(public_input.bitmask, domain());
@@ -96,7 +103,8 @@ contract BasicVerifier {
         });
 
         validate_evaluations(proof, evaluations_with_bitmask, challenges, evals_at_zeta);
-        Bw6Fr[] memory constraint_polynomial_evals = evaluations_with_bitmask.evaluate_constraint_polynomials(public_input.apk, evals_at_zeta);
+        Bw6Fr[] memory constraint_polynomial_evals =
+            evaluations_with_bitmask.evaluate_constraint_polynomials(public_input.apk, evals_at_zeta);
         Bw6Fr memory w = constraint_polynomial_evals.horner_field(challenges.phi);
         return (proof.r_zeta_omega.add(w)).eq(proof.q_zeta.mul(evals_at_zeta.vanishing_polynomial));
     }
@@ -104,7 +112,7 @@ contract BasicVerifier {
     function restore_challenges(
         AccountablePublicInput calldata public_input,
         SimpleProof calldata proof,
-        uint batch_size
+        uint256 batch_size
     ) internal pure returns (Challenges memory challenges) {
         // round 1
         // r: 316184119047170985678859850286058587173
@@ -136,9 +144,7 @@ contract BasicVerifier {
         // Reconstruct the commitment to the linearization polynomial using the commitments to the registers from the proof.
         // linearization polynomial commitment
         Bw6G1 memory r_comm = protocol.restore_commitment_to_linearization_polynomial(
-            challenges.phi,
-            evals_at_zeta.zeta_minus_omega_inv,
-            proof.register_commitments
+            challenges.phi, evals_at_zeta.zeta_minus_omega_inv, proof.register_commitments
         );
 
         // Aggregate the commitments to be opened in \zeta, using the challenge \nu.
@@ -156,15 +162,12 @@ contract BasicVerifier {
         register_evals[2] = proof.register_evaluations.partial_sums[0];
         register_evals[3] = proof.register_evaluations.partial_sums[1];
         register_evals[4] = proof.q_zeta;
-        (Bw6G1 memory w_comm, Bw6Fr memory w_at_zeta) = commitments.aggregate_claims_multiexp(register_evals, challenges.nus);
+        (Bw6G1 memory w_comm, Bw6Fr memory w_at_zeta) =
+            commitments.aggregate_claims_multiexp(register_evals, challenges.nus);
 
         // batched KZG openning
-        KzgOpening memory opening_at_zeta = KzgOpening({
-            c: w_comm,
-            x: challenges.zeta,
-            y: w_at_zeta,
-            proof: proof.w_at_zeta_proof
-        });
+        KzgOpening memory opening_at_zeta =
+            KzgOpening({c: w_comm, x: challenges.zeta, y: w_at_zeta, proof: proof.w_at_zeta_proof});
         KzgOpening memory opening_at_zeta_omega = KzgOpening({
             c: r_comm,
             x: evals_at_zeta.zeta_omega,
