+++ src/common/bls12377/Fp.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.17;
 import "../bw6761/Fr.sol";
 
 struct Bls12Fp {
-    uint a;
-    uint b;
+    uint256 a;
+    uint256 b;
 }
 
 library BLS12FP {
