+++ src/common/bls12377/G1.sol
@@ -9,12 +9,14 @@ struct Bls12G1 {
 
 library BLS12G1 {
     // BLS12_377_G1ADD
-    uint private constant G1_ADD = 0x13;
+    uint256 private constant G1_ADD = 0x13;
 
     function generator() internal pure returns (Bls12G1 memory) {
         return Bls12G1({
             x: Bls12Fp(0x8848defe740a67c8fc6225bf87ff54, 0x85951e2caa9d41bb188282c8bd37cb5cd5481512ffcd394eeab9b16eb21be9ef),
-            y: Bls12Fp(0x1914a69c5102eff1f674f5d30afeec4, 0xbd7fb348ca3e52d96d182ad44fb82305c2fe3d3634a9591afd82de55559c8ea6)
+            y: Bls12Fp(
+                0x1914a69c5102eff1f674f5d30afeec4, 0xbd7fb348ca3e52d96d182ad44fb82305c2fe3d3634a9591afd82de55559c8ea6
+                )
         });
     }
 
@@ -26,14 +28,11 @@ library BLS12G1 {
     }
 
     function point_in_g1_complement() internal pure returns (Bls12G1 memory) {
-        return Bls12G1({
-            x: Bls12Fp(0, 0),
-            y: Bls12Fp(0, 1)
-        });
+        return Bls12G1({x: Bls12Fp(0, 0), y: Bls12Fp(0, 1)});
     }
 
     function add(Bls12G1 memory p, Bls12G1 memory q) internal view returns (Bls12G1 memory) {
-        uint[8] memory input;
+        uint256[8] memory input;
         input[0] = p.x.a;
         input[1] = p.x.b;
         input[2] = p.y.a;
@@ -42,7 +41,7 @@ library BLS12G1 {
         input[5] = q.x.b;
         input[6] = q.y.a;
         input[7] = q.y.b;
-        uint[4] memory output;
+        uint256[4] memory output;
 
         assembly ("memory-safe") {
             if iszero(staticcall(gas(), G1_ADD, input, 256, output, 128)) {
@@ -55,7 +54,7 @@ library BLS12G1 {
         return from(output);
     }
 
-    function from(uint[4] memory x) internal pure returns (Bls12G1 memory) {
+    function from(uint256[4] memory x) internal pure returns (Bls12G1 memory) {
         return Bls12G1(Bls12Fp(x[0], x[1]), Bls12Fp(x[2], x[3]));
     }
 }
