+++ src/common/bw6761/Fp.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.17;
 import "../math/Math.sol";
 
 struct Bw6Fp {
-    uint a;
-    uint b;
-    uint c;
+    uint256 a;
+    uint256 b;
+    uint256 c;
 }
 
 library BW6FP {
@@ -16,7 +16,8 @@ library BW6FP {
         return Bw6Fp(
             0x122e824fb83ce0ad187c94004faff3eb926186a81d14688528275ef8087be41,
             0x707ba638e584e91903cebaff25b423048689c8ed12f9fd9071dcd3dc73ebff2e,
-            0x98a116c25667a8f8160cf8aeeaf0a437e6913e6870000082f49d00000000008b);
+            0x98a116c25667a8f8160cf8aeeaf0a437e6913e6870000082f49d00000000008b
+        );
     }
 
     function zero() internal pure returns (Bw6Fp memory) {
@@ -40,11 +41,7 @@ library BW6FP {
     }
 
     function gt(Bw6Fp memory x, Bw6Fp memory y) internal pure returns (bool) {
-        return (
-            x.a > y.a ||
-            (x.a == y.a && x.b > y.b) ||
-            (x.a == y.a && x.b == y.b && x.c > x.c)
-       );
+        return (x.a > y.a || (x.a == y.a && x.b > y.b) || (x.a == y.a && x.b == y.b && x.c > x.c));
     }
 
     function neg(Bw6Fp memory self) internal pure {
@@ -58,7 +55,7 @@ library BW6FP {
             uint8 carry = 0;
             (carry, z.c) = x.b.adc(y.c, carry);
             (carry, z.b) = x.b.adc(y.b, carry);
-            (     , z.a) = x.a.adc(y.a, carry);
+            (, z.a) = x.a.adc(y.a, carry);
         }
     }
 
@@ -82,7 +79,7 @@ library BW6FP {
             uint8 borrow = 0;
             (borrow, z.c) = m.b.sbb(y.c, borrow);
             (borrow, z.b) = m.b.sbb(y.b, borrow);
-            (      , z.a) = m.a.sbb(y.a, borrow);
+            (, z.a) = m.a.sbb(y.a, borrow);
         }
     }
 
@@ -91,14 +88,14 @@ library BW6FP {
         bytes32 a = bytes32(x.a);
         bytes32 b = bytes32(x.b);
         bytes32 c = bytes32(x.c);
-        for (uint i = 0; i < 32; i++) {
-            r[i] = c[31-i];
+        for (uint256 i = 0; i < 32; i++) {
+            r[i] = c[31 - i];
         }
-        for (uint j = 0; j < 32; j++) {
-            r[32+j] = b[31-j];
+        for (uint256 j = 0; j < 32; j++) {
+            r[32 + j] = b[31 - j];
         }
-        for (uint k = 0; k < 32; k++) {
-            r[64+k] = a[31-k];
+        for (uint256 k = 0; k < 32; k++) {
+            r[64 + k] = a[31 - k];
         }
         return r;
     }
